Multiprocessing fix, which needs to separate the main of generate_bc_trajectories into a separate file.
